was significantly harder
ran into a major issue: the nested directories could have duplicate names, so you need to store the unique path instead of just the dir_name itself. because of that, a lot of the directories in dir_sizes exceeded 1e7 (duplicates in different directories weren't separated) and weren't added to the final sum, and my initial answer was lower than the correct one.